Change some function
Add a return statement for some function.

Because of how the code is written, an error
is thrown if none of the if statements are
satisfied.

This cause an error to be thrown even though
the command is valid. The return statement helps
to prevent the error part to be thrown.