Add some assertions
Add some assertions.

The assertions check if there are any weird
anomalies. That is to say there must be
things we expect to be in a certain format,
or with a certain value must hold true.
Otherwise something is wrong with the
program for unknown reasons or
something is wrong with the user's device.