Change some details
Change some naming of the files and variables
to be more in line with the coding standard.

Most of the change was due to the pull request
review where reviewers have helped to state
some improvements to the naming conventions
used.

Also, I have changed the structure of the files
and removed some unnecessary variables or
functions.