Change user guide
The README is now updated with all information
that is critical to allow the user to understand all
the major features the chatbot offers. Anyone can
freely read it now

Add user guide

Add user guide

The README is now updated with all information
that is critical to allow the user to understand all
the major features the chatbot offers. Anyone can
freely read it now