Fix some bugs
Fix some bugs.

In the deletion of task, the task number
may not be valid. There may not be a check
used to ensure proper input. I have not added
a check to tell the user to reinput the correct
command.

Also, the remove command is also added.
This is to ensure people know about the
existence of a remove command.